services/cost-optimization-engine/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toolkit-cost-optimization-engine"
version = "0.1.0"
description = "Cost Optimization Engine: AI cost optimization and resource management with intelligent scaling and budget tracking."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "Toolkit Contributors" }]
keywords = ["cost-optimization", "resource-management", "budgeting", "scaling", "efficiency"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Typing :: Typed",
]
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.2.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "numpy>=1.24.0",
  "boto3>=1.29.0",
  "prometheus-client>=0.19.0",
]

[project.scripts]
toolkit-cost-optimizer = "toolkit_cost_optimization_engine.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.25.0",
  "ruff>=0.6.0",
  "pyright>=1.1.380",
  "rich>=13.0.0",
]
monitoring = [
  "prometheus-client>=0.19.0",
  "opentelemetry-api>=1.21.0",
  "opentelemetry-sdk>=1.21.0",
  "opentelemetry-instrumentation-fastapi>=0.42b0",
]
security = [
  "pip-audit>=2.7.0",
  "bandit>=1.7.5",
]
testing = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.25.0",
  "factory-boy>=3.3.0",
  "faker>=20.1.0",
]
release = [
  "build>=1.2.0",
  "twine>=5.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["toolkit_cost_optimization_engine"]
omit = [
  "*/toolkit_cost_optimization_engine/__main__.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 70

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "A", "COM"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S105", "S106", "S110", "S311", "B017"]
"src/toolkit_cost_optimization_engine/main.py" = ["B008"]
"src/toolkit_cost_optimization_engine/api/*" = ["B008"]
"src/toolkit_cost_optimization_engine/core/config.py" = ["S104", "S105"]

[tool.ruff.lint.isort]
known-first-party = ["toolkit_cost_optimization_engine"]

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
toolkit_cost_optimization_engine = ["py.typed"]

